requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
redis==5.0.4
google-api-python-client
youtube-transcript-api